Add taiga.service (Systemd compatibility)
This is one of three scripts created by my colleague to allow Systemd compatibility. It is not guaranteed to work, and because I'm not much of a programmer and am just filling in for him (he lacks the time to publish things at the moment), I can't guarantee it's in the right folder or includes enough comments. I'm just trying to help.